refactor: consolidate language-specific librarian-yaml updaters and tests
Unifies the language-specific LibrarianYamlUpdater implementations (Go, Java, Node, Python) into a single, shared LibrarianYamlUpdater class.
- Consolidates matching logic:
  - Handles single-version matching (by library name, component, or output directory).
  - Handles multi-version matching (using versionsMap for Java).
- Updates all release strategies and strategy tests to import from the unified updater.
- Merges all language-specific updater test files into a single, consolidated test suite.
- Removes duplicate and obsolete updater and test files.
